app/requirements.txt
ultralytics==8.0.196
opencv-python
pillow
gradio
networkx
numpy
pandas
matplotlib
